[Afterbuy] Bugfix of GetAfterbuyOrder
eBay Variant has different Mapping element "BaseProductData"
in the XML of Afterbuy Response. Therefore ALL eBay Variants
is mapped to the Master Product instead of the real Variant Product.

Hereby this bug gets fixed. But in the future it should be
more reasonable to retrieve the whole Order from E2 instead
of Afterbuy.

Signed-off-by: Xiaobo Sun <xiaobo.sun@telekom.de>